Delegate loading/processing view to `SearchPage`.
