+++ graylog2-web-interface/src/views/pages/NewSearchPage.tsx
@@ -16,12 +16,9 @@
  */
 import * as React from 'react';
 
-import { Spinner } from 'components/common';
-import useProcessHooksForView from 'views/logic/views/UseProcessHooksForView';
 import useCreateSavedSearch from 'views/logic/views/UseCreateSavedSearch';
 import normalizeSearchURLQueryParams from 'views/logic/NormalizeSearchURLQueryParams';
 import useQuery from 'routing/useQuery';
-import useLoadView from 'views/pages/useLoadView';
 
 import SearchPage from './SearchPage';
 
@@ -29,18 +26,8 @@ const NewSearchPage = () => {
   const query = useQuery();
   const { timeRange, queryString } = normalizeSearchURLQueryParams(query);
   const view = useCreateSavedSearch(undefined, timeRange, queryString);
-  useLoadView(view);
-  const [loaded, HookComponent] = useProcessHooksForView(view, query);
 
-  if (HookComponent) {
-    return HookComponent;
-  }
-
-  if (!loaded) {
-    return <Spinner />;
-  }
-
-  return <SearchPage />;
+  return <SearchPage view={view} />;
 };
 
 export default NewSearchPage;
